The current implementation breaks for `Dict`, `Set`, etc due to the specific implementation of #10, e.g.
```
julia> @testset ExtendedTestSet "begin" begin
           @test Dict(1 => 2) == Dict(2 => 2)
       end

=====================================================
begin: Test Failed at REPL[5]:2
  Expression: Dict(1 => 2) == Dict(2 => 2)
   Evaluated: Dict(1 => 2) == Dict(2 => 2)

<redacted>...
```

This PR will attempt to fix comparing of `Dict`, `Set`, etc while keeping the rest of the vectors and so on working.
```
julia> @testset ExtendedTestSet "begin" begin
           @test Dict(1 => 2) == Dict(2 => 2)
       end

=====================================================
begin: Test Failed
  Expression: Dict(1 => 2) == Dict(2 => 2)

Diff:
Dict(
-    1 => 2,
+    2 => 2,
)

<redacted>...
```

We achieved this by reverting to the implementation before the change. That is, to just evaluate regardless of what the LHS and RHS are in a `@test LHS == RHS` statement. There is one con, which is that `deepdiff` currently doesn't compare everything sufficiently well yet, especially if the objects are not part of the Julia base. This however should be an improvement that we will want to delegate to [DeepDiffs.jl](https://github.com/ssfrr/DeepDiffs.jl), which is good as that gives us a new good reason to help make improvement to that project 💪